Split test and production code
