+++ test/pyproject.toml
@@ -14,11 +14,15 @@ description = "Toga test"
 icon = "src/toga_test/resources/toga_test"
 sources = [
     'src/toga_test',
-    'src/tests',
+]
+test_sources = [
+    'tests',
 ]
 requires = [
     '../core',
     '../dummy',
+]
+test_requires = [
     'pytest==7.2.0',
     'pytest-asyncio==0.20.2',
 ]
@@ -55,7 +59,7 @@ flatpak_runtime_version = '42'
 flatpak_sdk = 'org.gnome.Sdk'
 
 [tool.briefcase.app.toga-test.windows]
-sources = [
+test_sources = [
     '../winforms/test_probes',
 ]
 requires = [
